requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
iniconfig==1.1.1
jsonpath==0.82
Markdown==3.3.4
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
